requirements.txt
pyserial==3.5
ttkbootstrap==1.8.0
Pillow==8.3.1